Show error stack traces on the same card as the log.
Expand in place like OpenUI Lang stream rows, with the trace and a Copy button on the card instead of a separate drill-in view.
